requirements.txt
aiogram==3.13.1
apscheduler==3.10.4
Pillow==10.3.0
requests==2.32.4
pytz==2024.2